Rename @vanilla addon to @vanilla-addon/vanilla
