+++ applications/vanilla/src/scripts/redux/reducer.ts
@@ -3,7 +3,7 @@
  * @license GPL-2.0-only
  */
 
-import { categoriesReducer } from "@vanilla/categories/categoriesReducer";
+import { categoriesReducer } from "@vanilla/addon-vanilla/categories/categoriesReducer";
 import { combineReducers } from "redux";
 import { IForumState } from "./state";
