+++ tsconfig-legacy.json
@@ -4,7 +4,7 @@
         "baseUrl": ".",
         "paths": {
             "@library/*": ["./library/src/scripts/*"],
-            "@vanilla/*": ["./applications/vanilla/src/scripts/*"],
+            "@vanilla/addon-vanilla/*": ["./applications/vanilla/src/scripts/*"],
             "@dashboard/*": ["./applications/dashboard/src/scripts/*"],
             "@rich-editor/*": ["./plugins/rich-editor/src/scripts/*"],
             "@knowledge/*": ["./plugins/knowledge/src/scripts/*"],
